implement checks on the frontend

+++ resources/js/Pages/Students/Index.svelte
@@ -31,13 +31,15 @@
                     </div>
 
                     <div class="mt-4 sm:mt-0 sm:ml-16 sm:flex-none">
-                        <Link
-                            href={route("students.create")}
-                            as="button"
-                            class="inline-flex items-center justify-center rounded-md border border-transparent bg-indigo-600 px-4 py-2 text-sm font-medium text-white shadow-sm hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2 sm:w-auto"
-                        >
-                            Add Student
-                        </Link>
+                        {#if $page.props.can.student_create}
+                            <Link
+                                href={route("students.create")}
+                                as="button"
+                                class="inline-flex items-center justify-center rounded-md border border-transparent bg-indigo-600 px-4 py-2 text-sm font-medium text-white shadow-sm hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-indigo-500 focus:ring-offset-2 sm:w-auto"
+                            >
+                                Add Student
+                            </Link>
+                        {/if}
                     </div>
                 </div>
                 <div class="mt-8 flex flex-col">
@@ -149,25 +151,28 @@
                                                 <td
                                                     class="relative whitespace-nowrap py-4 pl-3 pr-4 text-right text-sm font-medium sm:pr-6"
                                                 >
-                                                    <Link
-                                                        href={route(
-                                                            "students.edit",
-                                                            student.id
-                                                        )}
-                                                        class="text-indigo-600 hover:text-indigo-900"
-                                                    >
-                                                        Edit
-                                                    </Link>
-
-                                                    <button
-                                                        on:click={() =>
-                                                            handleDelete(
+                                                    {#if $page.props.can.student_edit}
+                                                        <Link
+                                                            href={route(
+                                                                "students.edit",
                                                                 student.id
                                                             )}
-                                                        class="ml-2 text-indigo-600 hover:text-indigo-900"
-                                                    >
-                                                        Delete
-                                                    </button>
+                                                            class="text-indigo-600 hover:text-indigo-900"
+                                                        >
+                                                            Edit
+                                                        </Link>
+                                                    {/if}
+                                                    {#if $page.props.can.student_delete}
+                                                        <button
+                                                            on:click={() =>
+                                                                handleDelete(
+                                                                    student.id
+                                                                )}
+                                                            class="ml-2 text-indigo-600 hover:text-indigo-900"
+                                                        >
+                                                            Delete
+                                                        </button>
+                                                    {/if}
                                                 </td>
                                             </tr>
                                         {/each}
